+++ package.json
@@ -35,41 +35,45 @@
   },
   "homepage": "https://github.com/flauc/angular2-notifications",
   "devDependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
+    "@angular-devkit/build-angular": "~0.12.0",
+    "@angular-devkit/build-ng-packagr": "~0.12.0",
+    "@angular/animations": "~9.0.0-next.12",
+    "@angular/common": "~9.0.0-next.12",
+    "@angular/compiler": "~9.0.0-next.12",
+    "@angular/compiler-cli": "~9.0.0-next.12",
+    "@angular/core": "~9.0.0-next.12",
+    "@angular/platform-browser": "~9.0.0-next.12",
+    "@angular/platform-browser-dynamic": "~9.0.0-next.12",
     "@compodoc/compodoc": "^1.1.7",
     "@types/jasmine": "^2.8.9",
     "@types/node": "^10.12.0",
     "angular-in-memory-web-api": "^0.6.1",
-    "codelyzer": "^4.5.0",
-    "core-js": "^2.5.7",
+    "codelyzer": "^5.1.2",
+    "core-js": "^2.5.4",
     "del": "^3.0.0",
     "gulp": "3.9.1",
     "gulp-rename": "^1.4.0",
     "gulp-rollup": "^2.16.2",
-    "jasmine-core": "2.9.1",
-    "jasmine-spec-reporter": "3.2.0",
-    "karma": "^3.1.3",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "0.2.0",
-    "karma-jasmine": "1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.1.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^4.2.0",
     "node-watch": "0.5.2",
     "protractor": "^5.4.1",
     "rollup": "^0.66.6",
     "run-sequence": "^2.2.1",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.3",
     "ts-node": "^7.0.1",
     "tsickle": "^0.33.0",
+    "tslib": "^1.9.0",
     "tslint": "^5.11.0",
-    "typescript": "^3.1.3",
-    "zone.js": "^0.8.26"
+    "typescript": "~3.6.4",
+    "zone.js": "~0.10.2"
   },
   "engines": {
     "node": ">=6.0.0"
