+++ src/lib/components/simple-notifications/simple-notifications.component.ts
