+++ src/lib/consts/default-icons.const.ts
