+++ src/lib/enums/notification-animation-type.enum.ts
