+++ src/lib/interfaces/icons.ts
