+++ src/lib/interfaces/notification-event.type.ts
