+++ src/lib/interfaces/notification.type.ts
