+++ src/lib/interfaces/options.type.ts
