--- src/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "angular2-notifications",
-  "version": "2.0.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/flauc/angular2-notifications"
-  },
-  "author": {
-    "name": "Filip Lauc",
-    "email": "filip.lauc93@gmail.com"
-  },
-  "keywords": [
-    "angular"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/flauc/angular2-notifications/issues"
-  },
-  "main": "angular2-notifications.umd.js",
-  "module": "angular2-notifications.js",
-  "jsnext:main": "angular2-notifications.js",
-  "typings": "angular2-notifications.d.ts",
-  "peerDependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
-  }
-}
