+++ src/public_api.ts
@@ -1,0 +1,11 @@
+export * from './lib/components/notification/notification.component';
+export * from './lib/components/simple-notifications/simple-notifications.component';
+export * from './lib/enums/notification-animation-type.enum';
+export * from './lib/enums/notification-type.enum';
+export * from './lib/interfaces/icons';
+export * from './lib/interfaces/notification-event.type';
+export * from './lib/interfaces/notification.type';
+export * from './lib/interfaces/options.type';
+export * from './lib/services/notifications.service';
+export * from './lib/simple-notifications.module';
+
